feat: refactor hooks and imports for better organization and clarity
