+++ src/features/auth/ResetPassword.tsx
@@ -8,7 +8,7 @@ import useAuthStore from '../../stores/useAuthStore';
 import { useResetPasswordMutation } from '../../hooks/useUserOperations';
 import { resetPasswordSchema } from './loginSchema';
 import ConfirmDialog from './ConfirmDialog';
-import useClearErrorMessage from '../../hooks/useClearAuthErrorMessage';
+import useClearErrorMessage from '../../hooks/useAuthError';
 
 // Material UI 元件
 import Box from '@mui/material/Box';
