+++ src/features/auth/VerifyCode.tsx
@@ -11,7 +11,7 @@ import {
   useResendVerificationCodeMutation,
   useVerifyCodeMutation,
 } from '../../hooks/useUserOperations';
-import useClearErrorMessage from '../../hooks/useClearAuthErrorMessage';
+import useClearErrorMessage from '../../hooks/useAuthError';
 
 // Material UI 元件
 import Box from '@mui/material/Box';
