+++ src/hooks/useProductOperations.ts
